Redo lights
