--- examples/07_TrackingLine.py
@@ -1,28 +1,0 @@
-import time
-import argparse
-from gpiozero import InputDevice
-
-line_pin_left = 22
-line_pin_middle = 27
-line_pin_right = 17
-
-left = InputDevice(pin=line_pin_right)
-middle = InputDevice(pin=line_pin_middle)
-right = InputDevice(pin=line_pin_left)
-
-def run():
-    status_right = right.value
-    status_middle = middle.value
-    status_left = left.value
-    print('left: %d   middle: %d   right: %d' %(status_right,status_middle,status_left))
-
-
-if __name__ == '__main__':
-    try:
-      while 1:
-        run()
-        time.sleep(0.3)
-    except KeyboardInterrupt:
-        pass
-
-
